+++ src/test/kotlin/in/kkkev/jjidea/contract/StubCommandExecutor.kt
@@ -188,6 +188,15 @@ class StubCommandExecutor(private val stub: JjStub) : CommandExecutor {
         revision: Revision
     ): CommandExecutor.CommandResult = TODO("Not needed for integration tests")
 
+    override fun fileTrack(filePaths: List<FilePath>): CommandExecutor.CommandResult =
+        TODO("Not needed for integration tests")
+
+    override fun fileUntrack(filePaths: List<FilePath>): CommandExecutor.CommandResult =
+        TODO("Not needed for integration tests")
+
+    override fun fileList(filePaths: List<FilePath>): CommandExecutor.CommandResult =
+        TODO("Not needed for integration tests")
+
     override fun resolveList(revision: Revision): CommandExecutor.CommandResult =
         TODO("Not needed for integration tests")